+++ src/styleguide/EightPxGrid.js
@@ -18,7 +18,7 @@ const GridBase = styled.div`
   }
 `;
 
-const Grid = ({ children }) => <GridBase size={getBodyFontSize() / 4}>{children}</GridBase>;
-Grid.propTypes = { children: PropTypes.node };
+const EightPxGrid = ({ children }) => <GridBase size={getBodyFontSize() / 4}>{children}</GridBase>;
+EightPxGrid.propTypes = { children: PropTypes.node };
 
-export default Grid;
+export default EightPxGrid;